+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   syntax, and general consistency improvements to elref parsing throughout.
   (Previously this syntax was only available in connectors).
 
+- Fixed: `auto-style-mode="inline"` multiple `style` attribute generation.
+
 - Breaking changes (CLI):
 
   - the input file was previously a positional argument, it now uses the `-i`